feat(modularize): separate out route
